requirements.txt
langchain
streamlit
faiss-cpu
langchain-community
langchain-openai
python-dotenv